backend/licimar_mvp_app/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-CORS==5.0.0
Flask-JWT-Extended==4.6.0
SQLAlchemy==2.0.40
cryptography==36.0.2
bcrypt==4.2.1
python-dotenv==1.0.1
marshmallow==3.23.2
webargs==8.6.0
Pillow==10.4.0
reportlab==4.2.5
pandas==2.2.3
openpyxl==3.1.5
python-dateutil==2.9.0
email-validator==2.2.0
gunicorn==21.2.0
